refactor: [GSW-2058] extract common gas fee calculation logic into utility function
