+++ packages/web/src/hooks/common/use-network-fee.ts
@@ -8,7 +8,8 @@ import { GasToken } from "@common/values/token-constant";
 import { useGnoswapContext } from "./use-gnoswap-context";
 import { makeEstimateGasTransaction } from "@utils/transaction-utils";
 import { WalletState } from "@states/index";
-import { DEFAULT_GAS_WANTED, GAS_WANTED_BUFFER_SAFE_MARGIN } from "@common/values";
+import { GAS_WANTED_BUFFER_SAFE_MARGIN } from "@common/values";
+import { calculateAdjustedGasFee } from "@utils/gas-utils";
 
 export interface UseNetworkFeeReturn {
   currentGasInfo: GasInfo | null;
@@ -101,18 +102,17 @@ export const useNetworkFee = (document: Document | null, gasInfo?: GasInfo | nul
         };
       }
 
-      const adjustGasUsedBN = BigNumber(resultGasUsed.gasUsed || DEFAULT_GAS_WANTED).multipliedBy(
+      const { gasFee, gasUsed, gasWanted } = calculateAdjustedGasFee(
+        resultGasUsed.gasUsed,
+        gasPrice,
         GAS_WANTED_BUFFER_SAFE_MARGIN,
       );
-      const adjustGasUsed = adjustGasUsedBN.toFixed(0, BigNumber.ROUND_DOWN);
-      const adjustGasPriceBN = BigNumber(gasPrice);
-      const gasFee = adjustGasPriceBN.multipliedBy(adjustGasUsed).toFixed(0, BigNumber.ROUND_UP);
 
       const gasInfoResult: GasInfo = {
-        gasFee: Number(gasFee),
-        gasUsed: Number(adjustGasUsed),
-        gasWanted: Number(adjustGasUsed),
-        gasPrice: gasPrice,
+        gasFee,
+        gasUsed,
+        gasWanted,
+        gasPrice,
         hasError: resultGasUsed.errorMessage !== null,
         simulateErrorMessage: resultGasUsed.errorMessage,
       };
