+++ packages/web/src/hooks/gas/use-get-estimate-gas-info.ts
@@ -1,4 +1,3 @@
-import BigNumber from "bignumber.js";
 import { useQuery, UseQueryOptions, UseQueryResult } from "@tanstack/react-query";
 
 import { useGnoswapContext } from "@hooks/common/use-gnoswap-context";
@@ -8,7 +7,8 @@ import { makeEstimateGasTransaction } from "@utils/transaction-utils";
 import { Document } from "src/types/transaction-messages.types";
 import { GasInfo } from "@hooks/gas";
 import { QUERY_KEY } from "@query/query-keys";
-import { DEFAULT_GAS_WANTED, GAS_WANTED_BUFFER_MULTIPLIER } from "@common/values";
+import { GAS_WANTED_BUFFER_MULTIPLIER } from "@common/values";
+import { calculateAdjustedGasFee } from "@utils/gas-utils";
 
 const REFETCH_INTERVAL = 5_000;
 
@@ -53,18 +53,17 @@ export const useGetEstimateGasInfo = (
         };
       }
 
-      const adjustGasUsedBN = BigNumber(resultGasUsed.gasUsed || DEFAULT_GAS_WANTED).multipliedBy(
+      const { gasFee, gasUsed, gasWanted } = calculateAdjustedGasFee(
+        resultGasUsed.gasUsed,
+        gasPrice,
         GAS_WANTED_BUFFER_MULTIPLIER,
       );
-      const adjustGasUsed = adjustGasUsedBN.toFixed(0, BigNumber.ROUND_DOWN);
-      const adjustGasPriceBN = BigNumber(gasPrice);
-      const gasFee = adjustGasPriceBN.multipliedBy(adjustGasUsed).toFixed(0, BigNumber.ROUND_UP);
 
       return {
-        gasFee: Number(gasFee),
-        gasUsed: Number(adjustGasUsed),
-        gasWanted: Number(adjustGasUsed),
-        gasPrice: gasPrice,
+        gasFee,
+        gasUsed,
+        gasWanted,
+        gasPrice,
         hasError: resultGasUsed.errorMessage !== null,
         simulateErrorMessage: resultGasUsed.errorMessage,
       };
